Move iterate_all_parameters to general utils
